requirements.txt
pygame==2.3.0
pycrypto==2.6.1